requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pillow==9.4.0
pytest==7.1.3
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
